udpate interfaces

+++ pkg/core/edge.go
@@ -1,15 +1,8 @@
 package core
 
-// type Edge struct {
-// 	source *Node
-// 	target *Node
-// 	weight float32
-// 	delay  int
-// }
-
 type Edge interface {
 	GetTarget() Node
 	GetSource() Node
-	GetWeight() float32
+	GetWeight() float64
 	GetDelay() int
 }
